+++ src/core/subsystem/systems/SSGraphics.h
@@ -11,7 +11,7 @@ public:
 	~SSGraphics();
 
 	virtual void Startup();
-	virtual void Update(const float deltaTime);
+	virtual void Update(const double deltaTime);
 	virtual void Shutdown();
 private:
 	GraphicsEngine* m_Graphics;
